requirements.txt
# Crunchy-Neck-Agent dependencies
# Python >= 3.10 required (dataclasses kw_only, match/case)

# ── exec / process tools ──────────────────────────────────────────────────────
# No third-party deps — stdlib only.

# ── comm_channels / ping_user ─────────────────────────────────────────────────
# No third-party deps for HTTP — stdlib urllib.request.

# ── shared env loading ────────────────────────────────────────────────────────
python-dotenv               # tools/env_loader.py — load .env into os.environ

# ── memory / remember tool ────────────────────────────────────────────────────
chromadb==0.4.24
onnxruntime==1.24.3
numpy<2.0

# ── snapshot tool ─────────────────────────────────────────────────────────────
Pillow                      # ImageGrab for desktop screenshots; also used by image_gen

# ── image_gen tool ────────────────────────────────────────────────────────────
google-genai                # Gemini image generation SDK

# ── skill_use / prompt builder ────────────────────────────────────────────────
pyyaml                      # YAML frontmatter parsing in skill_use.py

# ── memory_compaction / agent loop ────────────────────────────────────────────
openai                      # GPT-5.2 compaction calls (memory_compaction.py)
tiktoken                    # accurate token counting in estimate_tokens(); falls back to char/4 heuristic if absent

# ── computer_agent (Scout) ────────────────────────────────────────────────────
pyautogui                   # CUA action execution (click, type, key, scroll, drag)
pyperclip                   # clipboard-based typing for non-ASCII text (em dashes, smart quotes, etc.)
httpx                       # Chrome DevTools readiness poll in browser.py

# ── Future tools (add as implemented) ─────────────────────────────────────────
# anthropic               # Claude API
# slack-sdk               # Slack comm-channel adapter
# discord.py              # Discord comm-channel adapter
# pydantic                # schema validation
# nanoid                  # session ID generation (or use uuid, already stdlib)
